% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framinghamRiskEquation.R
\name{framingham_riskequation}
\alias{framingham_riskequation}
\title{Framingham Risk Equation}
\usage{
framingham_riskequation(df, years = 5, outcome = "CVD")
}
\arguments{
\item{df}{a dataframe, which must include the following columns/fields:

\itemize{
\item InternalID (numeric) case identifier
\item CardiovascularDisease (logical - TRUE or FALSE)
\item Diabetes (logical)
\item SmokingStatus (character) - "Smoker" if a smoker. All other values ignored
\item UrineAlbuminDate (character), UrineAlbuminValue (double), UrineAlbuminUnit (character)
\item PersistentProteinuria (logical)
\item eGFRDate (date), eGFRValue (double), eGFRUnits (character)
\item FamilialHypercholesterolaemia (logical)
\item LVH (logical) = left ventricular hyp0ertrophy
\item Cholesterol (double) in millimoles per litre (mmol/L or mM)
\item Triglycerides (double) in millimoles per litre (mmol/L or mM)
\item CholHDLRatio (double)
\item BP (character, two numbers separated by "/") in mmHg
\item Sex (character) "Female" or not
\item Ethnicity (character) "Aboriginal", "Torres Strait Islander", "Aboriginal/Torres Strait Islander" or not
\item Age (double)
}}

\item{years}{number of years to predict (from 4 to 12). default is 5 years.}

\item{outcome}{(default is "CVD")

 "CHD" - coronary heart disease. includes myocardial
 infarction, death from coronary heart disease plus angina pectoris and
 coronary insufficiency.

 "CVD" cardiovascular disease (the default) includes coronary heart disease,
 stroke (including transient ischaemia), congestive heart failure and
 peripheral vascular disease.}
}
\value{
a dataframe
\itemize{
 \item InternalID - the internalID of the person's row in the df (dataframe) parameter
 \item frisk - numeric (a number, or 'NA' if not enough information to computer),
 \item friskHI - either 'NA' or '>15\%'. >15\%' are groups considered equivalent risk to already having ischaemic heart disease.
}
}
\description{
Calculate cardiovascular disease risk according to Framingham Risk Equation
}
\details{
Equations and examples sourced from:

\itemize{
\item {"Guidelines for the management of Absolute cardiovascular disease risk. 2012"}
  by the National Vascular Disease Prevention Alliance of Australia.
  http://www.cvdcheck.org.au/pdf/Absolute_CVD_Risk_Full_Guidelines.pdf
\item {"Cardiovascular Disease Risk Profiles"}
  by Keaven M. Anderson, Patricia M. Odell, Peter W.F. Wilson, William B. Kannel
  in 'American Heart Journal' 1991; 121:293-298
\item {"An Updated Coronary Risk Profile - A Statement for Health Professionals"}
  by Keaven M. Anderson, Peter W.F. Wilson, Patricia M. Odell, William B. Kannel
  in AHA (American Heart Association) Medical/Scientific Statement, sourced from http://ahajournals.org
}
}
\examples{

framingham_riskequation(data.frame(InternalID = 1, BP = "135/80", Sex = "Female",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 230/48, Diabetes = TRUE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 5.96, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "Cardiovascular disease risk profiles" (Anderson 1991)
    # the worked answer in the paper is 0.22. this function returns 0.2189125
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.180
    #  (cvdcheck.org.au reports 18\%)
    # the same risk-factors except Sex = "Male" with outcome = "CVD" and years = 5 returns 0.202
    #  (cvdcheck.org.au reports 20\%)

framingham_riskequation(data.frame(InternalID = 2, BP = "130/80", Sex = "Male",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 240/45, Diabetes = FALSE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 6.22, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "An Updated Coronary Risk Profile" (Anderson 1991)
    # the worked answer in the paper is 0.192, this function returns 0.1919
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.133
    #  (cvdcheck.org.au reports 13\%)
    # the same risk-factors except LVH = TRUE (outcome = CVD, years = 5) returns 0.211
    #  (cvdcheck.org.au reports 21\%)

framingham_riskequation(data.frame(InternalID = 3, BP = "130/80", Sex = "Female",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 240/45, Diabetes = FALSE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 6.22, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "An Updated Coronary Risk Profile" (Anderson 1991)
    # the worked answer in the paper is 0.135, this function returns 0.1349
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.088
    #  (cvdcheck.org.au reports 9\%)
    # the same risk-factors except LVH = TRUE (outcome = CVD, years = 5) returns 0.150
    #  (cvdcheck.org.au reports 15\%)

}
